# src/main/java/grw/qut/inventory/domain/Manufacturer.java
package grw.qut.inventory.domain;

import lombok.Data;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;
import javax.persistence.Table;
import javax.persistence.UniqueConstraint;

@Data
@Entity
@Table(name = "manufacturer")
public class Manufacturer {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    @Column(name = "id", nullable = false)
    private Long id;

    @Column(name = "name", unique=true, nullable = false)
    private String name;

    @Column(name = "home_page")
    private String homePage;

    @Column(name = "phone")
    private String phone;
}